moved preferred colors test suite to its own test file
